feat(autofix): Add support for multiline input (#89973)
Switch all user inputs in Autofix to be TextArea's so users can use
shift + enter to add multiline text. No visible changes by default
though.

Also add some truncation logic to insight cards to show just the first
line of multiline cards by default, and show the rest when expanded.
